src/extract-team-images/pyproject.toml
[tool.poetry]
name = "extract-team-images"
version = "0.1.0"
description = ""
authors = ["Pravin Dahal <pravin.dahal@optimeering.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic-ai = "^0.0.22"
beautifulsoup4 = "^4.13.3"
pandas = "^2.2.3"
gdown = "^5.2.0"
httpx = "^0.28.1"
openpyxl = "^3.1.5"
pillow = "^11.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
